requirements.txt
aiofiles==0.5.0
asgiref==3.2.10
click==7.1.2
Django==3.1.2
fastapi==0.61.1
h11==0.11.0
passlib==1.7.4
pydantic==1.6.1
PyJWT==1.7.1
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
starlette==0.13.6
uvicorn==0.12.1
